More modifications to the reminders and waiting list
I noticed that the subject line fills with utf-8 tags when the number of characters exceeds 20 in the sms posts.  There is likely a work around for this 20 character limit but my quick solution is to go with the 20 limit for the subject.  I also separated out the sms messages from the html to facilitate this different subject line.